requirements.txt
customtkinter
Pillow
qrcode
keyboard
winshell